server/requirements.txt
absl-py==0.4.0
astor==0.7.1
bcrypt==3.1.4
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
click==6.7
cycler==0.10.0
Flask==1.0.2
flask-marshmallow==0.9.0
flask-notifyAll==1.0.3
Flask-SQLAlchemy==2.3.2
gast==0.2.0
grpcio==1.14.1
h5py==2.8.0
idna==2.7
IMAPClient==2.1.0
imbalanced-learn==0.3.3
imblearn==0.0
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
marshmallow==2.15.4
marshmallow-sqlalchemy==0.14.1
Pillow==5.2.0
protobuf==3.6.1
pycparser==2.18
PyJWT==1.6.4
pyparsing==2.2.0
pyserial==3.4
PySocks==1.6.8
python-dateutil==2.7.3
python-dotenv==0.9.1
pytz==2018.5
PyYAML==3.13
requests==2.19.1
shawk==0.4
sinchsms==1.0.4
six==1.11.0
sms==0.4
SQLAlchemy==1.2.10
termcolor==1.1.0
twilio==6.3.0
urllib3==1.23
Werkzeug==0.14.1
wincertstore==0.2
